:truck: Moved tree.rs units tests in src/tests/tree_tests.rs
