+++ src/tests/mod.rs
@@ -1,2 +1,3 @@
 mod ai_tests;
 mod evaluation_tests;
+mod tree_tests;
